requirements.txt
pymongo>=3.12.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0


